Restructured the lib to live without json in the future.
