+++ app/gamelib/src/game_state/components/mod.rs
@@ -1,0 +1,1 @@
+pub mod motion;
