+++ app/gamelib/src/game_state/components/motion.rs
@@ -1,0 +1,13 @@
+#[derive(Debug)]
+pub struct Position {
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
+}
+
+#[derive(Debug)]
+pub struct Velocity {
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
+}
